+++ src/plugins/sftp.ts
@@ -347,6 +347,9 @@ class SFtpPlugin extends deploy_objects.DeployPluginWithContextBase<SFTPContext>
                         globals: me.context.globals(),
                         mode: deploy_contracts.DataTransformerMode.Transform,
                         options: ctx.dataTransformerOptions,
+                        replaceWithValues: (val) => {
+                            return me.context.replaceWithValues(val);
+                        },
                         require: function(id) {
                             return me.context.require(id);
                         },